+++ nym-api/Cargo.toml
@@ -16,17 +16,12 @@ async-trait = { workspace = true }
 bs58 = { workspace = true }
 bip39 = { workspace = true }
 bincode.workspace = true
+console-subscriber = { workspace = true, optional = true } # nym-api needs to be built with RUSTFLAGS="--cfg tokio_unstable"
 cfg-if = { workspace = true }
 clap = { workspace = true, features = ["cargo", "derive", "env"] }
-console-subscriber = { workspace = true, optional = true } # validator-api needs to be built with RUSTFLAGS="--cfg tokio_unstable"
 dashmap = { workspace = true }
-dirs = { workspace = true }
 futures = { workspace = true }
-itertools = { workspace = true }
 humantime-serde = { workspace = true }
-k256 = { workspace = true, features = [
-    "ecdsa-core",
-] } # needed for the Verifier trait; pull whatever version is used by other dependencies
 moka = { workspace = true }
 pin-project = { workspace = true }
 rand = { workspace = true }
@@ -51,7 +46,6 @@ tendermint = { workspace = true }
 ts-rs = { workspace = true, optional = true }
 
 anyhow = { workspace = true }
-getset = { workspace = true }
 
 sqlx = { workspace = true, features = [
     "runtime-tokio-rustls",
@@ -66,30 +60,19 @@ zeroize = { workspace = true }
 
 # for axum server
 axum = { workspace = true, features = ["tokio"] }
-axum-extra = { workspace = true, features = ["typed-header"] }
 tower-http = { workspace = true, features = ["cors", "trace", "compression-br", "compression-deflate", "compression-gzip", "compression-zstd"] }
 utoipa = { workspace = true, features = ["axum_extras", "time"] }
 utoipauto = { workspace = true }
 utoipa-swagger-ui = { workspace = true, features = ["axum"] }
 tracing = { workspace = true }
 
-## ephemera-specific
-#actix-web = "4"
-#array-bytes = "6.0.0"
-#chrono = { version = "0.4.24", default-features = false, features = ["clock"] }
-#futures-util = "0.3.25"
-#serde_derive = "1.0.149"
-#uuid = { version = "1.3.0", features = ["serde", "v4"] }
-
 ## internal
-#ephemera = { path = "../ephemera" }
 nym-bandwidth-controller = { path = "../common/bandwidth-controller" }
 nym-ecash-contract-common = { path = "../common/cosmwasm-smart-contracts/ecash-contract" }
 nym-ecash-time = { path = "../common/ecash-time", features = ["expiration"] }
 nym-coconut-dkg-common = { path = "../common/cosmwasm-smart-contracts/coconut-dkg" }
 nym-compact-ecash = { path = "../common/nym_offline_compact_ecash" }
 nym-credentials-interface = { path = "../common/credentials-interface" }
-#nym-ephemera-common = { path = "../common/cosmwasm-smart-contracts/ephemera" }
 nym-config = { path = "../common/config" }
 cosmwasm-std = { workspace = true }
 nym-credential-storage = { path = "../common/credential-storage", features = [
@@ -102,11 +85,8 @@ cw3 = { workspace = true }
 cw4 = { workspace = true }
 nym-dkg = { path = "../common/dkg", features = ["cw-types"] }
 nym-gateway-client = { path = "../common/client-libs/gateway-client" }
-nym-inclusion-probability = { path = "../common/inclusion-probability" }
 nym-mixnet-contract-common = { path = "../common/cosmwasm-smart-contracts/mixnet-contract", features = ["utoipa"] }
-nym-vesting-contract-common = { path = "../common/cosmwasm-smart-contracts/vesting-contract" }
 nym-contracts-common = { path = "../common/cosmwasm-smart-contracts/contracts-common", features = ["naive_float", "utoipa"] }
-nym-multisig-contract-common = { path = "../common/cosmwasm-smart-contracts/multisig-contract" }
 nym-sphinx = { path = "../common/nymsphinx" }
 nym-pemstore = { path = "../common/pemstore" }
 nym-task = { path = "../common/task" }
@@ -121,7 +101,8 @@ nym-http-api-common = { path = "../common/http-api-common", features = ["utoipa"
 nym-serde-helpers = { path = "../common/serde-helpers", features = ["date"] }
 nym-ticketbooks-merkle = { path = "../common/ticketbooks-merkle" }
 nym-statistics-common = { path = "../common/statistics" }
-chrono.workspace = true
+nym-ecash-signer-check = { path = "../common/ecash-signer-check" }
+
 
 [features]
 no-reward = []
