+++ nym-api/src/ecash/api_routes/handlers.rs
@@ -4,8 +4,10 @@
 use crate::ecash::api_routes::aggregation::aggregation_routes;
 use crate::ecash::api_routes::issued::issued_routes;
 use crate::ecash::api_routes::partial_signing::partial_signing_routes;
+use crate::ecash::api_routes::signer_status::signer_status;
 use crate::ecash::api_routes::spending::spending_routes;
 use crate::support::http::state::AppState;
+use axum::routing::get;
 use axum::Router;
 
 pub(crate) fn ecash_routes() -> Router<AppState> {
@@ -14,4 +16,5 @@ pub(crate) fn ecash_routes() -> Router<AppState> {
         .merge(issued_routes())
         .merge(partial_signing_routes())
         .merge(spending_routes())
+        .route("/signer-status", get(signer_status))
 }
