+++ nym-api/src/main.rs
@@ -8,7 +8,6 @@ use ::nym_config::defaults::setup_env;
 use clap::Parser;
 use mixnet_contract_cache::cache::MixnetContractCache;
 use node_status_api::NodeStatusCache;
-use nym_bin_common::logging::setup_tracing_logger;
 use support::nyxd;
 use tracing::{info, trace};
 
@@ -23,6 +22,7 @@ pub(crate) mod node_describe_cache;
 mod node_performance;
 pub(crate) mod node_status_api;
 pub(crate) mod nym_nodes;
+mod signers_cache;
 mod status;
 pub(crate) mod support;
 mod unstable_routes;
@@ -32,9 +32,9 @@ async fn main() -> Result<(), anyhow::Error> {
     cfg_if::cfg_if! {if #[cfg(feature = "console-subscriber")] {
         // instrument tokio console subscriber needs RUSTFLAGS="--cfg tokio_unstable" at build time
         console_subscriber::init();
+    } else {
+        nym_bin_common::logging::setup_tracing_logger();
     }}
-
-    setup_tracing_logger();
 
     info!("Starting nym api...");
