+++ nym-api/src/network/handlers.rs
@@ -3,13 +3,19 @@
 
 use crate::network::models::{ContractInformation, NetworkDetails};
 use crate::node_status_api::models::AxumResult;
+use crate::signers_cache::handlers::signers_routes;
+use crate::support::config::CHAIN_STALL_THRESHOLD;
 use crate::support::http::state::AppState;
 use axum::extract::{Query, State};
 use axum::Router;
-use nym_api_requests::models::ChainStatusResponse;
+use nym_api_requests::models::{
+    ChainBlocksStatusResponse, ChainBlocksStatusResponseBody, ChainStatus, ChainStatusResponse,
+};
+use nym_api_requests::signable::SignableMessageBody;
 use nym_contracts_common::ContractBuildInformation;
 use nym_http_api_common::{FormattedResponse, OutputParams};
 use std::collections::HashMap;
+use time::OffsetDateTime;
 use tower_http::compression::CompressionLayer;
 use utoipa::ToSchema;
 
@@ -17,18 +23,24 @@ pub(crate) fn nym_network_routes() -> Router<AppState> {
     Router::new()
         .route("/details", axum::routing::get(network_details))
         .route("/chain-status", axum::routing::get(chain_status))
+        .route(
+            "/chain-blocks-status",
+            axum::routing::get(chain_blocks_status),
+        )
         .route("/nym-contracts", axum::routing::get(nym_contracts))
         .route(
             "/nym-contracts-detailed",
             axum::routing::get(nym_contracts_detailed),
         )
+        .nest("/signers", signers_routes())
         .layer(CompressionLayer::new())
 }
 
 #[utoipa::path(
     tag = "network",
     get,
-    path = "/v1/network/details",
+    context_path = "/v1/network",
+    path = "/details",
     responses(
         (status = 200, content(
             (NetworkDetails = "application/json"),
@@ -50,7 +62,8 @@ async fn network_details(
 #[utoipa::path(
     tag = "network",
     get,
-    path = "/v1/network/chain-status",
+    context_path = "/v1/network",
+    path = "/chain-status",
     responses(
         (status = 200, content(
             (ChainStatusResponse = "application/json"),
@@ -79,6 +92,47 @@ async fn chain_status(
     }))
 }
 
+#[utoipa::path(
+    tag = "network",
+    get,
+    context_path = "/v1/network",
+    path = "/chain-blocks-status",
+    responses(
+        (status = 200, content(
+            (ChainBlocksStatusResponse = "application/json"),
+            (ChainBlocksStatusResponse = "application/yaml"),
+            (ChainBlocksStatusResponse = "application/bincode")
+        ))
+    ),
+    params(OutputParams)
+)]
+async fn chain_blocks_status(
+    Query(params): Query<OutputParams>,
+    State(state): State<AppState>,
+) -> FormattedResponse<ChainBlocksStatusResponse> {
+    let output = params.get_output();
+
+    let current_time = OffsetDateTime::now_utc();
+    let latest_cached_block = state
+        .chain_status_cache
+        .get_or_refresh(&state.nyxd_client)
+        .await
+        .ok();
+    let chain_status = latest_cached_block
+        .as_ref()
+        .map(|detailed| detailed.stall_status(current_time, CHAIN_STALL_THRESHOLD))
+        .unwrap_or(ChainStatus::Unknown);
+
+    output.to_response(
+        ChainBlocksStatusResponseBody {
+            current_time,
+            latest_cached_block,
+            chain_status,
+        }
+        .sign(state.private_signing_key()),
+    )
+}
+
 // it's used for schema generation so dead_code is fine
 #[allow(dead_code)]
 #[derive(ToSchema)]
@@ -90,7 +144,7 @@ pub(crate) struct ContractVersionSchemaResponse {
     /// version is any string that this implementation knows. It may be simple counter "1", "2".
     /// or semantic version on release tags "v0.7.0", or some custom feature flag list.
     /// the only code that needs to understand the version parsing is code that knows how to
-    /// migrate from the given contract (and is tied to it's implementation somehow)
+    /// migrate from the given contract (and is tied to its implementation somehow)
     pub version: String,
 }
 
@@ -104,7 +158,8 @@ pub struct ContractInformationContractVersion {
 #[utoipa::path(
     tag = "network",
     get,
-    path = "/v1/network/nym-contracts",
+    context_path = "/v1/network",
+    path = "/nym-contracts",
     responses(
         (status = 200, content(
             (HashMap<String, ContractInformationContractVersion> = "application/json"),
@@ -151,7 +206,8 @@ pub struct ContractInformationBuildInformation {
 #[utoipa::path(
     tag = "network",
     get,
-    path = "/v1/network/nym-contracts-detailed",
+    context_path = "/v1/network",
+    path = "/nym-contracts-detailed",
     responses(
         (status = 200, content(
             (HashMap<String, ContractInformationBuildInformation> = "application/json"),
