+++ tools/internal/testnet-manager/src/manager/dkg_skip.rs
@@ -412,10 +412,10 @@ impl NetworkManager {
 
         let mut receivers = Vec::new();
         for signer in &ctx.ecash_signers {
-            // send 101nym to the admin
+            // send 250nym to the admin
             receivers.push((
                 signer.data.cosmos_account.address.clone(),
-                admin.mix_coins(101_000000),
+                admin.mix_coins(250_000000),
             ))
         }
